refactor(client): extract logic from register modal component
